refactor(progress): move to script setup
