+++ src/packages/__VUE/progress/index.taro.ts
@@ -1,0 +1,13 @@
+import Progress from './progress.taro.vue';
+import type { ComponentPublicInstance } from 'vue';
+import { withInstall } from '@/packages/utils';
+
+withInstall(Progress);
+
+export type { ProgressProps } from './progress.taro.vue';
+
+export type { ProgressSize, ProgressStatus } from './types';
+
+export type ProgressInstance = ComponentPublicInstance & InstanceType<typeof Progress>;
+
+export { Progress, Progress as default };
